+++ CartMS/Models/CartItemModel.cs
@@ -7,17 +7,17 @@ namespace CartMS.Models
     [PrimaryKey(nameof(customer_id), nameof(seller_id), nameof(product_id))]
     public class CartItemModel
 	{
-        public long customer_id { get; set; }
+        public int customer_id { get; set; }
 
-        public long seller_id { get; set; }
+        public int seller_id { get; set; }
 
-        public long product_id { get; set; }
+        public int product_id { get; set; }
 
         public string product_name { get; set; } = "";
 
-        public decimal unit_price { get; set; }
+        public float unit_price { get; set; }
 
-        public decimal freight_value { get; set; }
+        public float freight_value { get; set; }
 
         public int quantity { get; set; }
