+++ Common/Driver/TransactionMark.cs
@@ -1,4 +1,4 @@
 ﻿namespace Common.Driver
 {
-	public record TransactionMark(int tid, TransactionType type, long actorId);
+	public record TransactionMark(int tid, TransactionType type, int actorId);
 }
