+++ OrderMS.Common/Models/OrderItemModel.cs
@@ -16,38 +16,38 @@ namespace OrderMS.Common.Models
 
         /* another way (must remove the foreign key in ordermodel):
         [ForeignKey("order")]
-        public long order_id { get; set; }
+        public int order_id { get; set; }
         public OrderModel order { get; set; }
         */
-        public long order_id { get; set; }
+        public int order_id { get; set; }
 
-        public long order_item_id { get; set; }
+        public int order_item_id { get; set; }
 
-        public long product_id { get; set; }
+        public int product_id { get; set; }
 
         public string product_name { get; set; } = "";
 
-        public long seller_id { get; set; }
+        public int seller_id { get; set; }
 
-        public decimal unit_price { get; set; }
+        public float unit_price { get; set; }
 
         public DateTime shipping_limit_date { get; set; }
 
-        public decimal freight_value { get; set; }
+        public float freight_value { get; set; }
 
         // not present in olist
         public int quantity { get; set; }
 
         // without freight value
-        public decimal total_items { get; set; }
+        public float total_items { get; set; }
 
         // without freight value
-        public decimal total_amount { get; set; }
+        public float total_amount { get; set; }
 
         // can be derived from total_items - total_amount
         // incentive of item is not of concern to the order
         // the seller must compute 
-        // public decimal total_incentive { get; set; }
+        // public float total_incentive { get; set; }
 
         public OrderItemModel() { }
