+++ OrderMS.Common/Repositories/IOrderRepository.cs
@@ -6,8 +6,8 @@ namespace OrderMS.Common.Repositories
 	public interface IOrderRepository
 	{
         public IEnumerable<OrderModel> GetAll();
-        public IEnumerable<OrderModel> GetByCustomerId(long customerId);
-        public OrderModel? GetOrder(long orderId);
+        public IEnumerable<OrderModel> GetByCustomerId(int customerId);
+        public OrderModel? GetOrder(int orderId);
     }
 }
