+++ PaymentMS/Repositories/IPaymentRepository.cs
@@ -6,7 +6,7 @@ namespace PaymentMS.Repositories
 	public interface IPaymentRepository
 	{
 
-        IEnumerable<OrderPaymentModel> GetByOrderId(long orderId);
+        IEnumerable<OrderPaymentModel> GetByOrderId(int orderId);
 
     }
 }
